pyproject.toml
[project]
name = "rag"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "beautifulsoup4>=4.13.4",
    "requests>=2.32.3",
    "python-multipart>=0.0.20",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.45.0",
    "sentence-transformers>=5.2.3",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.126.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "rank-bm25>=0.2.2",
    "datasets>=4.2.0",
    "ragas>=0.3.8"
]
